+++ src/ReadOnlyModel.js
@@ -1,17 +1,17 @@
 'use strict';
 
 /*
-* Copyright © 2024-present, Ian Silverstone
-*
-* See the LICENSE file at the top-level directory of this distribution
-* for licensing information.
-*
-* Removal or modification of this copyright notice is prohibited.
-*/
+ * Copyright © 2024-present, Ian Silverstone
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ */
 
-import BaseModel from './BaseModel.js';
+import TableModel from './TableModel.js';
 
-class ReadOnlyModel extends BaseModel {
+class ReadOnlyModel extends TableModel {
   constructor(db, pgp, schema, logger) {
     super(db, pgp, schema, logger);
   }