+++ src/index.js
@@ -1,31 +1,25 @@
 'use strict';
 
 /*
-* Copyright © 2024-present, Ian Silverstone
-*
-* See the LICENSE file at the top-level directory of this distribution
-* for licensing information.
-*
-* Removal or modification of this copyright notice is prohibited.
-*/
+ * Copyright © 2024-present, Ian Silverstone
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ */
 
 import DB from './DB.js';
-import BaseModel from './BaseModel.js';
+import TableModel from './TableModel.js';
 import ReadOnlyModel from './ReadOnlyModel.js';
-import * as schemaBuilder  from './utils/schemaBuilder.js';
+import * as schemaBuilder from './utils/schemaBuilder.js';
 import * as validation from './utils/validation.js';
 
-export {
-  DB,
-  BaseModel,
-  ReadOnlyModel,
-  schemaBuilder,
-  validation,
-};
+export { DB, TableModel, ReadOnlyModel, schemaBuilder, validation };
 
 export default {
   DB,
-  BaseModel,
+  TableModel,
   ReadOnlyModel,
   schemaBuilder,
   validation,