+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## [0.1.0-beta.1] - 2025-04-17
 
 ### Added
+
 - Initial beta release of `pg-schemata`.
 - Core features include:
-  - `BaseModel` class for CRUD operations and filtering.
+  - `TableModel` class for CRUD operations and filtering.
   - `DB` class for singleton database connection handling.
   - Schema definition utilities: `createColumnSet`, `addAuditFields`.
   - Input validation utilities.